+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## New Features
 
 - **_Hitbox-based Hitscan Collision_** setting
+- **Recoil-pitch scale percent** setting
 - **_'IDDET'_ cheat**, to find exits in the automap
+- **Number of total message lines** setting, to let the message review show more messages
 
 ## Changes
 
@@ -9,4 +11,5 @@ None.
 
 ## Bug Fixes
 
-None.
+- **Message review overriding message durations even when the new one were shorter**
+- **Message review ignoring the forced-duration setting**